+++ packages/hooks/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @0xsequence/hooks
 
+## 5.3.9
+
+### Patch Changes
+
+- - erc1155SaleContract fix
+  - X auth support for Embedded Wallet
+
+## 5.3.8
+
+### Patch Changes
+
+- erc1155SaleContract fix
+
 ## 5.3.7
 
 ### Patch Changes